refactor(): update to core 2.0.0
refactor(UIRouter): change UIRouter component props

remove history prop in favor of router and plugin ones. History
implementation are provided by the core as plugins and are exported by
the react version of the router. The router prop can be used to pass
your own instance if you want to manually bootstrap it.